+++ javax/src/main/java/me/panpf/javax/util/Formatx.java
@@ -323,7 +323,7 @@ public class Formatx {
      * Returns the total time of formatting that can be displayed
      */
     @NotNull
-    public static String totalTime(long totalTimeMillis, @NotNull FormatTotalTimeConfig config) {
+    public static String totalTime(long totalTimeMillis, @NotNull TotalTimeConfig config) {
         long finalTotalTimeMillis = totalTimeMillis >= 0 ? totalTimeMillis : 0;
         int finalLevel = Math.max(config.getLevel(), 0);
 
@@ -381,7 +381,7 @@ public class Formatx {
     public static String totalTime(long totalTimeMillis, int level, @NotNull String divider,
                                    @NotNull String daySuffix, @NotNull String hourSuffix, @NotNull String minuteSuffix,
                                    @NotNull String secondSuffix, @NotNull String millisecondSuffix) {
-        return totalTime(totalTimeMillis, new FormatTotalTimeConfig(level, divider, daySuffix, hourSuffix, minuteSuffix, secondSuffix, millisecondSuffix));
+        return totalTime(totalTimeMillis, new TotalTimeConfig(level, divider, daySuffix, hourSuffix, minuteSuffix, secondSuffix, millisecondSuffix));
     }
 
     /**
@@ -735,5 +735,68 @@ public class Formatx {
      */
     public static String durationShort(long duration) {
         return duration(duration, true);
+    }
+
+    public static class TotalTimeConfig {
+        /**
+         * level 0: Accurate to milliseconds; 1: Accurate to seconds; 2: Accurate to minute; 3: Accurate to hour; 4: Accurate to day; 5 or more always returns 0 seconds
+         */
+        private int level;
+        @NotNull
+        private String divider;
+        @NotNull
+        private String daySuffix;
+        @NotNull
+        private String hourSuffix;
+        @NotNull
+        private String minuteSuffix;
+        @NotNull
+        private String secondSuffix;
+        @NotNull
+        private String millisecondSuffix;
+
+        public TotalTimeConfig(int level, @NotNull String divider, @NotNull String daySuffix, @NotNull String hourSuffix, @NotNull String minuteSuffix, @NotNull String secondSuffix, @NotNull String millisecondSuffix) {
+            this.level = level;
+            this.divider = divider;
+            this.daySuffix = daySuffix;
+            this.hourSuffix = hourSuffix;
+            this.minuteSuffix = minuteSuffix;
+            this.secondSuffix = secondSuffix;
+            this.millisecondSuffix = millisecondSuffix;
+        }
+
+        public int getLevel() {
+            return level;
+        }
+
+        @NotNull
+        public String getDivider() {
+            return divider;
+        }
+
+        @NotNull
+        public String getDaySuffix() {
+            return daySuffix;
+        }
+
+        @NotNull
+        public String getHourSuffix() {
+            return hourSuffix;
+        }
+
+        @NotNull
+        public String getMinuteSuffix() {
+            return minuteSuffix;
+        }
+
+        @NotNull
+        public String getSecondSuffix() {
+            return secondSuffix;
+        }
+
+        @NotNull
+        public String getMillisecondSuffix() {
+            return millisecondSuffix;
+        }
     }
 }